+++ src/nespresso/bot/handlers/admin/register.py
@@ -1,11 +1,13 @@
 from aiogram import Dispatcher
 
-from nespresso.bot.handlers.admin.commands import logs, messages, send
+from nespresso.bot.handlers.admin.commands import admin, logs, messages, send, senda
 
 
 def RegisterAdminHandlers(dp: Dispatcher) -> None:
     dp.include_routers(
+        admin.router,
         logs.router,
         messages.router,
         send.router,
+        senda.router,
     )