+++ CHANGELOG.md
@@ -22,6 +22,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 ---
 
 ## [0.1.0] - 2025-05-25 - Vlad
+### Added
+- Admin's command to send messages to all registered users
+- Admin's command to see admin's commands
+
 ### Updated
 - Functionality of TgUser repository